FLAGS DONE.
I was having some trouble making this one in Illustrator so I pulled
out Inkscape & tried to figure out how to use it. It displays fine for
me in browser, but Illustrator renders it as a black mass.
I don’t know, I’m going to go ahead and complete the flag section
though.